backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary      
pyodbc               
python-dotenv        